Refactor: extracted Grammar quality check code into its own module
